+++ app/api/app.py
@@ -1,0 +1,5 @@
+from fastapi import FastAPI
+from app.api.schema import graphql_app
+
+app = FastAPI()
+app.include_router(graphql_app, prefix="/graphql")
