+++ app/database/engine.py
@@ -1,0 +1,5 @@
+from sqlmodel import SQLModel, create_engine
+
+engine = create_engine("sqlite:///database.db")
+
+SQLModel.metadata.create_all(engine)
